+++ src/palette.ts
@@ -5,7 +5,7 @@ All rights reserved.
  
 This code is licensed under the BSD-3 license found in the LICENSE file in the root directory of this source tree and at https://opensource.org/licenses/BSD-3-Clause.
 **/
-import { Error, Neutral, NeutralVariant, Orange, Primary, Purple, Success, Warning } from '../types/colorType';
+import { Error, Neutral, NeutralVariant, Orange, Primary, Purple, Success, Warning } from './colorType';
 
 /* Primary Colors / User Interface Colors */